requirements.txt
Django==4.1.3
SimpleWebSocketServer==0.1.2
sqlparse==0.4.2
influxdb_client==1.36.1
pusher==3.3.2
pysher==1.0.8
paho-mqtt==1.6.1
psycopg2
schedule
django-cors-headers>=3.14.0
django-rest-framework
# Plotting --------------------------------------------------------------------
pandas>=1.1.4
djangorestframework-jwt
google-generativeai
